Fixing OpenGL widget

+++ core/cameracalibration/photoStationCalibrator.h
@@ -66,7 +66,7 @@ private:
                 int idx = (*offset)[j];
                 auto& Qs = absoluteSetupLocation[j].orientation;
                 auto& Cs = absoluteSetupLocation[j].position;
-                for (size_t i = 0; i < N; ++i)
+                for (int i = 0; i < N; ++i)
                 {
                     if (!patternPoints[j][i].size())
                         continue;
